Fix all 63 mypy findings in solver/+ml/, add strict scoped CI job (#384)
Correctly scoped (mypy custom_components/nimbus_load/solver
custom_components/nimbus_load/ml --ignore-missing-imports
--follow-imports=silent), the tree had 63 real findings, not zero --
split out of #358 per Mark Purcell's own suggestion, since solver/ and
ml/ are pure numpy with zero HA imports, hence zero type-stub ambiguity
unlike the rest of the integration.

Two real categories, both fixed at the root rather than suppressed:

1. Optional-not-narrowed (Mark's own priority -- "the ones I would look
   at first," and the exact class of gap that let #372's float-vs-
   datetime mismatch through). Three real shapes:
   - A repeated subscript/attribute expression (`x[i]`, `trained.gbrt`)
     doesn't narrow across a None-check the way a plain local variable
     does -- fixed by binding to a local (walrus or plain assignment)
     first, everywhere this pattern appeared (ml/model.py's three
     duplicated build_features() call sites, GBRT.predict access,
     quantile-model access; ml/gbrt.py's tree recursion).
   - A boolean flag (has_export_bonus, has_quantile_models,
     use_early_stopping) doesn't carry the None-check it was computed
     from forward to a later, separate use of the original Optional --
     fixed with a documented `assert` restating the real invariant at
     each use site (solver/network.py x4, solver/stochastic.py x3,
     solver/p2p_export.py x2, ml/gbrt.py x1).
   - Two correlated-but-independently-typed Optional values (lower_kw/
     upper_kw in elements.py's confidence-band validator) -- fixed by
     checking both explicitly instead of relying on an XOR invariant
     mypy can't see across variables.
   - counterfactuals.py's tune_two_threshold() could genuinely return
     None past its own declared non-Optional signature (an all-
     identical-candidate grid search finds no valid pair) -- fixed with
     a real raise, not a suppression, matching this project's own
     "represent honestly, don't paper over" convention.

2. numpy stub dtype-widening (Mark's own lower-priority call --
   "mostly stub noise, not real bugs, confirm case by case"). Chained
   float64 array arithmetic (+, -, *, /) widens to floating[Any] in
   numpy's own stubs regardless of the real runtime dtype -- fixed with
   explicit .astype(np.float64)/dtype= casts at each site (network.py,
   quality_report.py, reference_benchmark.py), confirming case by case
   per Mark's own suggestion rather than a blanket per-module ignore.

Two real, precise type-annotation fixes alongside the above: p2p_export.py
and stochastic.py's own per-day bonus-cap grouping dicts were typed
dict[object, list[int]] (object has no .isoformat()) when the real key
type is always datetime.date (or date | None for stochastic.py's global-
fallback case) -- corrected, not suppressed.

New CI job (typecheck-solver-ml): strict, no `|| true`, same
--follow-imports=silent invocation, gates every future PR. No baseline
file -- the correctly-scoped count was small and real, not noise
requiring one. The existing whole-package mypy job stays advisory
(separate, much bigger #40 Platinum-quality-scale effort).

Verified: zero mypy errors across all 21 solver/+ml/ source files.
Every touched file's own full test suite re-run and passing (450
solver/network tests, 93 quality-report/battery-sign/contract-risk
tests, 23 stochastic tests, 16 gbrt tests, 7 reference-benchmark tests,
36 predict()/coordinator tests, 12 counterfactuals tests, 8 p2p-export-
dedup tests -- 645 tests total across the files this touched, zero
regressions). ruff check/format clean on every file.

+++ custom_components/nimbus_load/solver/reference_benchmark.py
@@ -226,8 +226,12 @@ def _reference_persistence_curves(
     matching the existing test_solver_forecast_regret.py fixture's own
     "yesterday, flat-shifted" reasoning.
     """
-    solar_persistence_kw = np.roll(solar_real_kw, 2) * 0.85
-    load_persistence_kw = np.roll(load_real_kw, 1) * 1.10
+    # mypy issue #384: numpy widens ndarray[float64] * python-float to
+    # floating[Any] in its own stubs -- a real stub-precision gap, not a
+    # real bug (see this project's own CLAUDE.md GBRT-vs-k-NN finding for
+    # the general pattern).
+    solar_persistence_kw = (np.roll(solar_real_kw, 2) * 0.85).astype(np.float64)
+    load_persistence_kw = (np.roll(load_real_kw, 1) * 1.10).astype(np.float64)
     return solar_persistence_kw, load_persistence_kw
